+++ resources/views/revisions/partials/sidebar.blade.php
@@ -72,9 +72,7 @@
                     {{-- D2, point 1: how many revised entities this group holds. --}}
                     <x-badge>{{ $group->entities->count() }}</x-badge>
                 </span>
-                <svg class="h-4 w-4 fill-current transition-transform" :class="{ '-rotate-90': ! (filter.trim() !== '' || open) }" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-                    <path fill-rule="evenodd" d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" clip-rule="evenodd" />
-                </svg>
+                <x-tabler-chevron-down class="h-4 w-4 transition-transform" x-bind:class="{ '-rotate-90': ! (filter.trim() !== '' || open) }" />
             </button>
 
             {{-- Force-open while filtering so matches are always visible. --}}